launch: maintain directory watch
Maintain a Dirwatch object together with a configuration file. We treat
it as associated with the configuration and recreate it with each
reload request. On reload-failure, we keep the old Dirwatch.

This does not hook the dirwatch into the configuration parser, but
merely makes the launcher maintain the dirwatch object and react to
events.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>